add policies for admin panel
Only show admin panels for admin user (for now)

+++ app/Http/Controllers/Back/AdminAlbumController.php
@@ -25,9 +25,11 @@ class AdminAlbumController extends Controller
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index()
     {
+        $this->authorize('index', Album::class);
         $albums = Album::with('pictures')->latest()->get();
 
         return view('admin.albums.index', [
@@ -39,9 +41,11 @@ class AdminAlbumController extends Controller
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
+        $this->authorize('create', Album::class);
         return view('admin.albums.create');
     }
 
@@ -50,17 +54,21 @@ class AdminAlbumController extends Controller
      *
      * @param PictureUploadRequest $request
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(PictureUploadRequest $request)
     {
         //TODO Store categories
-        $album = Album::create([
-            'title' => $request->input('title'),
-            'body' => $request->input('body'),
-            'publish' => $request->input('publish'),
-            'user_id' => Auth::id(),
-            'password' => Hash::make($request->input('password')),
-        ]);
+        $album = new Album();
+        $album->title = $request->input('title');
+        $album->body = $request->input('body');
+        $album->publish = $request->input('publish');
+        $album->user_id = Auth::id();
+        $album->password = Hash::make($request->input('password'));
+
+        $this->authorize('store', $album);
+
+        $album->save();
 
         foreach ($request->pictures as $uploaderPicture) {
             $picture = new Picture();
@@ -76,9 +84,11 @@ class AdminAlbumController extends Controller
      *
      * @param Album $album
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show(Album $album)
     {
+        $this->authorize('show', Album::class);
         return view('admin.albums.show', ['album' => $album]);
     }
 
@@ -87,9 +97,11 @@ class AdminAlbumController extends Controller
      *
      * @param Album $album
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Album $album)
     {
+        $this->authorize('edit', $album);
         return view('admin.albums.edit', ['album' => $album]);
     }
 
@@ -99,6 +111,7 @@ class AdminAlbumController extends Controller
      * @param PictureUploadRequest $request
      * @param $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(PictureUploadRequest $request, $id)
     {
@@ -108,8 +121,11 @@ class AdminAlbumController extends Controller
         $album->title = $request->input('title');
         $album->body = $request->input('body', '');
         $album->publish = $request->input('publish', false);
+        //TODO Keep current ?
         $album->user_id = Auth::id();
         $album->password = Hash::make($request->input('password'));
+
+        $this->authorize('update', $album);
 
         $album->save();
 
@@ -131,6 +147,8 @@ class AdminAlbumController extends Controller
      */
     public function destroy(Album $album)
     {
+        $this->authorize('destroy', $album);
+
         // Suppression des fichiers (dossier)
         Storage::disk('uploads')->deleteDirectory('albums/' . $album->id);
         $album->pictures()->delete();
